+++ kernel/include/fs/fs.h
@@ -158,8 +158,8 @@ int write_char(dev_t dev, const char *buf, int count);
 /* system calls */
 int do_open(const char *pathname, int flags, mode_t mode);
 int do_close(int fd);
-int do_read(int fd, char *buf, int count);
-int do_write(int fd, const char *buf, int count);
+ssize_t do_read(int fd, char *buf, int count);
+ssize_t do_write(int fd, const char *buf, int count);
 off_t do_lseek(int fd, off_t offset, int whence);
 int do_stat(const char *filename, struct stat_t *statbuf);
 int do_access(const char *filename);
